Implement concrete ClockTempRecorder

+++ src/common/type/StateParam.java
@@ -2,12 +2,12 @@ package common.type;
 
 // Parameter structure to pass down subject's states to observers
 public class StateParam {
-    private float temperature; // in celcius
+    private double temperature; // in celcius
     // More field can be added if the functionality gets upgraded and handle different facdtors like humid and e.t.c
-    public StateParam(int temperature){
+    public StateParam(double temperature){
         this.temperature = temperature;
     }
-    public int getTemperature() {
+    public double getTemperature() {
         return this.temperature;
     }
 }
